Handle sub-millisecond response time
In the case where a ping returns a latency of less than one millisecond the status should color-grade according to the provided configuration.

When the ping command is used against a hostname to which latency is very low e.g. localhost or default route the current implementation will consider the ping failed.

When the latency is less than a millisecond the ping command will return a measurement in the form of `time<1ms`. Currently, the [regex](https://github.com/soluml/daskeyboard-applet--dasping/blob/b78d990c6bd64a19af69410721e1876f47a82b31/index.js#L5) used to extract the latency measurement assumes that the latency will always be in the form of `time=xms`. Therefore, when the latency is very, very low the regex will not match and is considered a failure.

```
> ping 8.8.8.8 -n 1

Pinging 8.8.8.8 with 32 bytes of data:
Reply from 8.8.8.8: bytes=32 time=46ms TTL=115

Ping statistics for 8.8.8.8:
    Packets: Sent = 1, Received = 1, Lost = 0 (0% loss),
Approximate round trip times in milli-seconds:
    Minimum = 46ms, Maximum = 46ms, Average = 46ms
```

```
> ping 127.0.0.1 -n 1

Pinging 127.0.0.1 with 32 bytes of data:
Reply from 127.0.0.1: bytes=32 time<1ms TTL=128

Ping statistics for 127.0.0.1:
    Packets: Sent = 1, Received = 1, Lost = 0 (0% loss),
Approximate round trip times in milli-seconds:
    Minimum = 0ms, Maximum = 0ms, Average = 0ms
```

Update the regex to match both `time=` and `time<` as proposed in this pull request.

After writing this up I can see that perspective given my example of using `127.0.0.1`. It may seem weird that one would ping their local machine for a status when the status is shown on a device that one would be using on that machine when the status changes.

However, my actual use case is to ping two hosts:
 1. a known external ip (8.8.8.8)
 2. a known internal ip

Since my 'known internal ip' happens to be my primary network router - and as such is the default route on my machine - the latency recorded is very little. When I setup the monitoring I immediately realized that my local router would also show a failure status.

 - I hand tested these changes using Das Keyboard Q version v3.3.3 on a Windows 11 machine.
 - Added some unit test cases.

 - I added a `.prettierrc.yaml` file with values that closely match the existing patterns observed in this project in order to minimize the noise produced in this pull request by the [default prettier config values](https://prettier.io/docs/en/options.html). My editor is setup to always run prettier when I touch a file. Having the `.prettierrc.yaml` will ensure that all future prettier runs will format the code in the same way and reduce PR noise from others that may use prettier.